chore: move withFromOrXpub into caller
